initial set of changes to brute forcing
