+++ amass/brute.go
@@ -12,17 +12,30 @@ import (
 	"github.com/miekg/dns"
 )
 
+// BruteForceQueryTypes contains the DNS record types that service queries for.
+var BruteForceQueryTypes = []string{
+	"TXT",
+	"CNAME",
+	"A",
+	"AAAA",
+}
+
 // BruteForceService is the Service that handles all brute force name generation
-// within the architecture. This is achieved by watching all the NEWSUB events.
+// within the architecture.
 type BruteForceService struct {
 	core.BaseService
 
+	max    utils.Semaphore
 	filter *utils.StringFilter
 }
 
 // NewBruteForceService returns he object initialized, but not yet started.
 func NewBruteForceService(config *core.Config, bus *core.EventBus) *BruteForceService {
-	bfs := &BruteForceService{filter: utils.NewStringFilter()}
+	num := (len(resolvers) * 100) / len(BruteForceQueryTypes)
+	bfs := &BruteForceService{
+		max:    utils.NewSimpleSemaphore(num),
+		filter: utils.NewStringFilter(),
+	}
 
 	bfs.BaseService = *core.NewBaseService(bfs, "Brute Forcing", config, bus)
 	return bfs
@@ -41,8 +54,8 @@ func (bfs *BruteForceService) OnStart() error {
 			}
 		}
 		go bfs.startRootDomains()
+		go bfs.processRequests()
 	}
-	go bfs.processRequests()
 	return nil
 }
 
@@ -62,17 +75,12 @@ func (bfs *BruteForceService) processRequests() {
 }
 
 func (bfs *BruteForceService) goodRequest(req *core.Request) bool {
-	if !bfs.Config().BruteForcing {
-		return false
-	}
-
 	if !bfs.Config().IsDomainInScope(req.Name) {
 		return false
 	}
 
-	bfs.SetActive()
-
 	var ok bool
+	bfs.SetActive()
 	for _, r := range req.Records {
 		t := uint16(r.Type)
 
@@ -93,34 +101,84 @@ func (bfs *BruteForceService) startRootDomains() {
 
 // NewSubdomain is called by the Name Service when proper subdomains are discovered.
 func (bfs *BruteForceService) NewSubdomain(req *core.Request, times int) {
-	if times >= bfs.Config().MinForRecursive {
+	if times == bfs.Config().MinForRecursive {
 		bfs.SendRequest(req)
 	}
 }
 
 func (bfs *BruteForceService) performBruteForcing(subdomain, domain string) {
-	if bfs.filter.Duplicate(subdomain) {
+	subdomain = strings.ToLower(subdomain)
+	domain = strings.ToLower(domain)
+	req := &core.Request{
+		Name:   subdomain,
+		Domain: domain,
+	}
+	if subdomain == "" || domain == "" || bfs.filter.Duplicate(subdomain) ||
+		GetWildcardType(req) == WildcardTypeDynamic {
 		return
 	}
 
+	curIdx := 0
+	maxIdx := 7
+	delays := []int{10, 25, 50, 75, 100, 150, 250, 500}
 	bfs.SetActive()
 	t := time.NewTicker(time.Second)
 	defer t.Stop()
-	fire := time.NewTicker(100 * time.Microsecond)
-	defer fire.Stop()
 	for _, word := range bfs.Config().Wordlist {
 		select {
 		case <-bfs.Quit():
 			return
 		case <-t.C:
 			bfs.SetActive()
-		case <-fire.C:
-			bfs.Bus().Publish(core.NewNameTopic, &core.Request{
-				Name:   strings.ToLower(word + "." + subdomain),
-				Domain: domain,
-				Tag:    core.BRUTE,
-				Source: bfs.String(),
-			})
+		default:
+			if !bfs.max.TryAcquire(1) {
+				time.Sleep(time.Duration(delays[curIdx]) * time.Millisecond)
+				if curIdx < maxIdx {
+					curIdx++
+				}
+				continue
+			}
+			curIdx = 0
+			if word != "" {
+				name := strings.ToLower(word + "." + subdomain)
+				go bfs.bruteForceResolution(name, subdomain, domain)
+			}
 		}
 	}
+}
+
+func (bfs *BruteForceService) bruteForceResolution(name, sub, domain string) {
+	defer bfs.max.Release(1)
+	defer bfs.SetActive()
+
+	if name == "" || domain == "" {
+		return
+	}
+
+	var answers []core.DNSAnswer
+	for _, t := range BruteForceQueryTypes {
+		if a, err := Resolve(name, t); err == nil {
+			answers = append(answers, a...)
+			// Do not continue if a CNAME was discovered
+			if t == "CNAME" {
+				break
+			}
+		}
+	}
+
+	req := &core.Request{
+		Name:   name,
+		Domain: domain,
+	}
+	if len(answers) == 0 || MatchesWildcard(req) {
+		return
+	}
+
+	bfs.Bus().Publish(core.NameResolvedTopic, &core.Request{
+		Name:    name,
+		Domain:  domain,
+		Records: answers,
+		Tag:     core.BRUTE,
+		Source:  bfs.String(),
+	})
 }
